requirements.txt
pdfplumber>=0.10.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
reportlab>=4.0.0
